kernel: MFC the recent -Wswitch fixes.
From master:

1647c7bd435093709089283aa68ef831b34caef9
2c995f40849b20e79621e20576e89e785a0d92a7
f46681358e7500a0505796f5744110299cdc99d8

+++ sys/kern/subr_diskiocom.c
@@ -108,7 +108,7 @@ disk_iocom_uninit(struct disk *dp)
 }
 
 int
-disk_iocom_ioctl(struct disk *dp, int cmd, void *data)
+disk_iocom_ioctl(struct disk *dp, u_long cmd, void *data)
 {
 	struct file *fp;
 	struct disk_ioc_recluster *recl;
