+++ sys/net/ip_mroute/ip_mroute.h
@@ -379,7 +379,7 @@ struct sockopt;
 extern int	(*ip_mrouter_set)(struct socket *, struct sockopt *);
 extern int	(*ip_mrouter_get)(struct socket *, struct sockopt *);
 extern int	(*ip_mrouter_done)(void);
-extern int	(*mrt_ioctl)(int, caddr_t);
+extern int	(*mrt_ioctl)(u_long, caddr_t);
 
 #endif /* _KERNEL */
