+++ sys/netinet6/ip6_mroute.c
@@ -329,7 +329,7 @@ ip6_mrouter_get(struct socket *so, struct sockopt *sopt)
  * Handle ioctl commands to obtain information from the cache
  */
 int
-mrt6_ioctl(int cmd, caddr_t data)
+mrt6_ioctl(u_long cmd, caddr_t data)
 {
 	int error = 0;
